Add EpollQuicUtils class
Motivation:

We may be able to use advanced features if Epoll is used. Let's make it easier for people to use these advanced features

Modifications:

- Add EpollQuicUtils class which will provide with utility methods when the user uses the native epoll transport
- Make use of this class in the tests

Result:

Easier for users to make use of advanced features
