+++ pom.xml
@@ -868,6 +868,14 @@
       <groupId>io.netty</groupId>
       <artifactId>netty-transport-native-epoll</artifactId>
       <version>${netty.version}</version>
+      <scope>compile</scope>
+      <!-- Let's mark as optional as it is not strictly needed -->
+      <optional>true</optional>
+    </dependency>
+    <dependency>
+      <groupId>io.netty</groupId>
+      <artifactId>netty-transport-native-epoll</artifactId>
+      <version>${netty.version}</version>
       <classifier>linux-x86_64</classifier>
       <scope>test</scope>
     </dependency>
